requirements.txt
Django
django-filter       
markdown
djangorestframework
whitenoise
mysqlclient
